chore: extract list menu items builder into hook
